requirements.txt
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
colorama==0.4.6
decorator==5.1.1
Django==4.2
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==2.0.1
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pygments==2.17.2
PyJWT==2.8.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing-extensions==4.11.0
tzdata==2024.1
wcwidth==0.2.13
